refactor: move changes under deps-check/ into local/
